Revert "refactor: outsource useeffect"
