+++ src/components/UI/navigation/UserIcon.tsx
@@ -3,18 +3,15 @@ import {
     FC,
     SetStateAction,
     createContext,
+    useEffect,
     useRef,
     useState,
 } from 'react'
-import SettingsPopup from './SettingsPopup'
-import { UserProps } from '../../../types/UserProps'
 import { getImage } from '../../../helper/getImage'
-import { getRandomChars } from '../../../helper/getRandomChars'
+import { UserProps } from '../../../types/UserProps'
 import { getUser } from '../../../helper/getUser'
-import { useCloseUserIcon } from '../../../hooks/useCloseUserIcon'
-import { useGetAuthUrl } from '../../../hooks/useGetAuthUrl'
-import { useLogUserOut } from '../../../hooks/useLogUserOut'
-import { useSetRedirectUrl } from '../../../hooks/useSetRedirectUrl'
+import SettingsPopup from './SettingsPopup'
+import { getRandomChars } from '../../../helper/getRandomChars'
 
 export const ContextFilterLanguageExpanded = createContext<
     [boolean, Dispatch<SetStateAction<boolean>>] | undefined
@@ -64,7 +61,20 @@ export const UserIcon: FC<UserIconProps> = ({ anchorRef, buttonRef }) => {
         sessionStorage.setItem('twitch_logged_in', 'true')
     }
 
-    useLogUserOut(fetchUser, user)
+    useEffect(() => {
+        const isLoggedIn = sessionStorage.getItem('twitch_logged_in')
+        isLoggedIn === 'true' && !user && fetchUser()
+        const timer = setTimeout(
+            () =>
+                isLoggedIn === 'false' &&
+                (sessionStorage.removeItem('twitch_logged_in'),
+                sessionStorage.removeItem('twitch_access_state'),
+                sessionStorage.removeItem('twitch_access_token'),
+                sessionStorage.removeItem('twitch_user')),
+            0
+        )
+        return () => clearTimeout(timer)
+    }, [user])
 
     const handleButtonClick = () => {
         setDropdownActive((prev) => !prev)
@@ -88,17 +98,74 @@ export const UserIcon: FC<UserIconProps> = ({ anchorRef, buttonRef }) => {
         }
     }
 
-    useGetAuthUrl(setRedirectUrl, state)
-    useSetRedirectUrl(redirectUrl || '')
+    useEffect(() => {
+        const getAuthUrl = async () => {
+            if (state.length) {
+                try {
+                    const response = await fetch(
+                        `https://twitch-backend.vercel.app/api/auth-url?state=${state}`
+                    )
+                    const data = await response.json()
+                    setRedirectUrl(data.url)
+                } catch (error) {
+                    console.error('Error fetching auth URL:', error)
+                }
+            }
+        }
+        getAuthUrl()
+    }, [state])
 
-    useCloseUserIcon(
-        buttonRef,
-        dropdownActive,
-        filterLanguageExpanded,
-        popupRef,
-        setDropdownActive,
-        setFilterLanguageExpanded
-    )
+    useEffect(() => {
+        if (redirectUrl) {
+            window.location.href = redirectUrl
+        }
+    }, [redirectUrl])
+
+    useEffect(() => {
+        const handleClickOutside = (event: MouseEvent) => {
+            if (
+                popupRef.current &&
+                !popupRef.current.contains(event.target as Node) &&
+                buttonRef.current &&
+                !buttonRef.current.contains(event.target as Node)
+            ) {
+                setDropdownActive(false)
+                setFilterLanguageExpanded(false)
+            }
+        }
+
+        if (dropdownActive) {
+            document.addEventListener('mousedown', handleClickOutside)
+        } else {
+            document.removeEventListener('mousedown', handleClickOutside)
+        }
+
+        return () => {
+            document.removeEventListener('mousedown', handleClickOutside)
+        }
+    }, [buttonRef, dropdownActive])
+
+    useEffect(() => {
+        const handleEscape = (event: KeyboardEvent) => {
+            if (
+                popupRef.current &&
+                !popupRef.current.contains(event.target as Node) &&
+                !filterLanguageExpanded
+            ) {
+                setDropdownActive(false)
+            }
+        }
+
+        if (dropdownActive && !filterLanguageExpanded) {
+            document.addEventListener('keydown', handleEscape)
+        } else {
+            document.removeEventListener('keydown', handleEscape)
+        }
+
+        return () => {
+            document.removeEventListener('keydown', handleEscape)
+        }
+    }, [dropdownActive, filterLanguageExpanded])
 
     return (
         <>
