+++ src/components/streamFeed/StreamFeed.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useContext, useState } from 'react'
+import { useCallback, useContext, useEffect, useState } from 'react'
 import { StreamProps } from '../../types/StreamProps'
 
 import SkeletonFeed from '../skeleton/SkeletonFeed'
@@ -29,7 +29,6 @@ import { getItemFromStorage } from '../../helper/getItemFromStorage'
 import { getSearchFilter } from '../../helper/getSearchFilter'
 import { getStreams } from '../../helper/getStreams'
 import { setItemInStorage } from '../../helper/setItemInStorage'
-import { useLoadStreams } from '../../hooks/useLoadStreams'
 
 const bgColors = [
     'bg-gradient-to-tr from-red-400 to-red-800',
@@ -182,7 +181,16 @@ const StreamFeed = () => {
         setStreamData,
     ])
 
-    useLoadStreams(loadStreams)
+    useEffect(() => {
+        loadStreams()
+        const refresh = setInterval(() => {
+            loadStreams()
+        }, 120000)
+        return () => {
+            clearInterval(refresh)
+        }
+    }, [loadStreams])
+
     useFocusTrap(error, filteredStreamData)
 
     if (error) {
