--- src/hooks/useSetRedirectUrl.ts
@@ -1,9 +1,0 @@
-import { useEffect } from 'react'
-
-export const useSetRedirectUrl = (redirectUrl: string) => {
-    useEffect(() => {
-        if (redirectUrl) {
-            window.location.href = redirectUrl
-        }
-    }, [redirectUrl])
-}
